# Improve performance on `work_packages/update_ancestors_service_spec`